Extract ruleflow group tests into separate test class
